refactor: Pass isSeller prop to Products component in ProductsPage
The code changes in this commit update the ProductsPage component to pass the isSeller prop to the Products component. This prop is used to conditionally render certain elements in the ProductItem component based on whether the user is a seller or not. This change improves the flexibility and reusability of the Products component.